Update webpage structure and script
Webpage layout has been remodelled for better synchronization with the updated script. The button and former paragraphs have been replaced with a div section containing three paragraphs. The script was extensively overhauled to offer enhanced manipulation and investigation of class names on the first paragraph of the div section. This change is implemented in line with the project's updated specifications requiring more sophisticated handling of classes in scripts.