"Add Dashboard section and improve HTML formatting"
In this commit, a new section called Dashboard was added to the HTML body including an h3 tag. Moreover, unnecessary whitespace in the contents of the p tags has been removed to improve the HTML formatting. The JavaScript part of code has been extended with operations to create, append, and prepend new HTML elements to the body. Two methods, dashboard.after() and dashboard.before(), have been added to modify the position of created elements.