Add interactive elements and event handlers
An input field and a button were added to the jst.html file. The function of the button was then extended to include a "mouseover" and "mouseleave" event as well as a "double-click" event. The "mouseover" event changes the background color of the h1 element with the class of "dashboard", while the "mouseleave" event reverts the color back. The "double-click" event triggers an alert box. Note that previously written console.log and element selection lines were commented out.