Added interactive message feature
Added a new div section with class "messages" to display messages from Di. Enhanced the function of the button and input events. On button click, if the input value length is greater than zero, the input value will be displayed in the agentSpan. The button will remain disabled unless the input value exists. This commit improves user interaction by allowing users to send and display messages.