"Refactor HTML and move JavaScript code to separate file"
Refactored the HTML code in 'jst.html' by commenting out the old divs and adding a new form group and unordered list. The motive for this change was to clean up the clutter in the code and make it neater.
Also moved all JavaScript operations from 'jst.html' to a new file 'script.js' to separate concerns and make the code more maintainable, especially when it gets more complex. This also helps in improving the page load time, as the browser does not have to load the JavaScript in the HTML.