Add display functionality to todo list
A function 'showList' is added to the main.js file of the todo list application. This function uses the existing 'todoList' array to display each todo item in a list format in the user interface. Each todo item has buttons for 'Done', 'Important', and 'Delete'. We call 'showList' at the end to ensure the list is displayed as soon as the script is loaded.