Add input validation and importance feature to todo list
In this change, two new features were added to the Todo List. Firstly, the `<input>` in `index.html` was changed to require a value before submitting the form. This prevents users from adding empty todos.

Additional functionality was also added in `main.js` to mark a todo item as important. The importance status influences the color of the todo item title, helping the user to easily identify important tasks. This is achieved through adding an 'isImportant' property to each item and controlling its status with an onclick event.